requirements.txt
image
nltk
requests
unidecode
bs4
pytesseract
num2words
pillow
spacy